chore(aggregated-assets): move page merging to the platform layer
mergeAssetsDataPages was in the api package but the api never calls it - its
only consumers are useAssetsData and useStocksData, both hooks. Merging the
pages of an infinite query is a consumption decision, not something the api
does, so it belongs with the hooks in @features/platform-aggregated-assets.

It was also misplaced twice: its test was still in dada-client importing through
the shim, the same code-away-from-test split just corrected for chunkCurrencyIds
and deepMergeCryptoAssets.

Both move to features/platform/aggregated-assets as pagination.ts and
pagination.test.ts, assertions unchanged. The platform package gains
@domain/api-aggregated-assets for AssetsDataWithPagination, which is the correct
downward dependency; live-common's shim now re-exports from the platform
package.

Still 246 tests: 216 in dada-client, 17 in the api package, 13 here.
