+++ domain/api/aggregated-assets/src/index.ts
@@ -6,4 +6,3 @@ export * from "./accessors";
 export * from "./api";
 export * from "./errors";
 export * from "./market";
-export * from "./pagination";
