+++ features/platform/aggregated-assets/src/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./pagination";
