+++ features/platform/aggregated-assets/src/pagination.test.ts
@@ -1,5 +1,5 @@
-import { mergeAssetsDataPages } from "../mergeAssetsDataPages";
-import type { AssetsDataWithPagination } from "../../state-manager/types";
+import { mergeAssetsDataPages } from "./pagination";
+import type { AssetsDataWithPagination } from "@domain/api-aggregated-assets";
 
 const makePage = (overrides: Partial<AssetsDataWithPagination> = {}): AssetsDataWithPagination => ({
   cryptoAssets: {},
